feat: add role permission information

+++ app/Http/Controllers/Admin/InformationCategoryController.php
@@ -6,6 +6,7 @@ use Illuminate\Http\Request;
 use Yajra\DataTables\DataTables;
 use App\Models\InformationCategory;
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Auth;
 use App\Http\Requests\Admin\InformationCategoryRequest;
 
 class InformationCategoryController extends Controller
@@ -15,15 +16,18 @@ class InformationCategoryController extends Controller
      */
     public function index()
     {
+        if (!Auth::user()->can('Manage Informasi')) {
+            return redirect()->back()->with('error', 'Maaf, Anda tidak memiliki akses untuk halaman tersebut');
+        }
         if (request()->ajax()) {
-            $data = InformationCategory::latest()->get();
+            $data = InformationCategory::latest();
             return DataTables::of($data)
                 ->addColumn('action', function ($data) {
                     $actionEdit = route('information-category.edit', $data->id);
                     $actionDelete = route('information-category.destroy', $data->id);
                     return "<div class='d-flex justify-content-center'>" .
-                        view('components.action.edit', ['action' => $actionEdit]) .
-                        view('components.action.delete', ['action' => $actionDelete, 'id' => $data->id]) .
+                        view('components.action.edit', ['action' => $actionEdit, 'name' => 'Informasi']) . '&nbsp;' .
+                        view('components.action.delete', ['action' => $actionDelete, 'id' => $data->id, 'name' => 'Informasi']) .
                         "</div>";
                 })
                 ->rawColumns(['action'])
@@ -37,6 +41,9 @@ class InformationCategoryController extends Controller
      */
     public function create()
     {
+        if (!Auth::user()->can('Create Informasi')) {
+            return redirect()->back()->with('error', 'Maaf, Anda tidak memiliki akses untuk halaman tersebut');
+        }
         return view('admins.information-category.create-edit');
     }
 
@@ -45,6 +52,9 @@ class InformationCategoryController extends Controller
      */
     public function store(InformationCategoryRequest $request)
     {
+        if (!Auth::user()->can('Create Informasi')) {
+            return redirect()->back()->with('error', 'Maaf, Anda tidak memiliki akses untuk halaman tersebut');
+        }
         InformationCategory::create($request->validated());
         return redirect()->route('information-category.index')
             ->with('success', 'Kategori informasi berhasil ditambahkan');
@@ -63,6 +73,9 @@ class InformationCategoryController extends Controller
      */
     public function edit(InformationCategory $informationCategory)
     {
+        if (!Auth::user()->can('Edit Informasi')) {
+            return redirect()->back()->with('error', 'Maaf, Anda tidak memiliki akses untuk halaman tersebut');
+        }
         return view('admins.information-category.create-edit', compact('informationCategory'));
     }
 
@@ -71,6 +84,9 @@ class InformationCategoryController extends Controller
      */
     public function update(InformationCategoryRequest $request, InformationCategory $informationCategory)
     {
+        if (!Auth::user()->can('Edit Informasi')) {
+            return redirect()->back()->with('error', 'Maaf, Anda tidak memiliki akses untuk halaman tersebut');
+        }
         $informationCategory->update($request->validated());
         return redirect()->route('information-category.index')
             ->with('success', 'Kategori informasi berhasil diperbarui');
@@ -81,6 +97,9 @@ class InformationCategoryController extends Controller
      */
     public function destroy(InformationCategory $informationCategory)
     {
+        if (!Auth::user()->can('Delete Informasi')) {
+            return redirect()->back()->with('error', 'Maaf, Anda tidak memiliki akses untuk halaman tersebut');
+        }
         $informationCategory->delete();
         return redirect()->route('information-category.index')
             ->with('success', 'Kategori informasi berhasil dihapus');
