+++ app/Http/Controllers/Admin/InformationController.php
@@ -7,6 +7,7 @@ use Illuminate\Http\Request;
 use Yajra\DataTables\DataTables;
 use App\Models\InformationCategory;
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Auth;
 use App\Http\Requests\Admin\InformationRequest;
 
 class InformationController extends Controller
@@ -16,8 +17,11 @@ class InformationController extends Controller
      */
     public function index()
     {
+        if (!Auth::user()->can('Manage Informasi')) {
+            return redirect()->back()->with('error', 'Maaf, Anda tidak memiliki akses untuk halaman tersebut');
+        }
         if (request()->ajax()) {
-            $data = Information::with('informationCategory')->latest()->get();
+            $data = Information::with('informationCategory')->latest();
             return DataTables::of($data)
                 ->editColumn('image', function ($data) {
                     return "<img src='" . asset($data->image) . "' class='img-fluid' style='max-width: 100px'>";
@@ -26,8 +30,8 @@ class InformationController extends Controller
                     $actionEdit = route('information.edit', $data->id);
                     $actionDelete = route('information.destroy', $data->id);
                     return "<div class='d-flex justify-content-center'>" .
-                        view('components.action.edit', ['action' => $actionEdit]) .
-                        view('components.action.delete', ['action' => $actionDelete, 'id' => $data->id]) .
+                        view('components.action.edit', ['action' => $actionEdit, 'name' => 'Informasi']) . '&nbsp;' .
+                        view('components.action.delete', ['action' => $actionDelete, 'id' => $data->id, 'name' => 'Informasi']) .
                         "</div>";
                 })
                 ->rawColumns(['action', 'image'])
@@ -41,6 +45,9 @@ class InformationController extends Controller
      */
     public function create()
     {
+        if (!Auth::user()->can('Create Informasi')) {
+            return redirect()->back()->with('error', 'Maaf, Anda tidak memiliki akses untuk halaman tersebut');
+        }
         $informationCategories = InformationCategory::orderBy('name')->get();
         return view('admins.information.create-edit', compact('informationCategories'));
     }
@@ -50,6 +57,9 @@ class InformationController extends Controller
      */
     public function store(InformationRequest $request)
     {
+        if (!Auth::user()->can('Create Informasi')) {
+            return redirect()->back()->with('error', 'Maaf, Anda tidak memiliki akses untuk halaman tersebut');
+        }
         $data = $request->validated();
         if ($request->hasFile('image')) {
             $data['image'] = 'storage/' . $request->file('image')->store('images/information', 'public');
@@ -72,6 +82,9 @@ class InformationController extends Controller
      */
     public function edit(Information $information)
     {
+        if (!Auth::user()->can('Edit Informasi')) {
+            return redirect()->back()->with('error', 'Maaf, Anda tidak memiliki akses untuk halaman tersebut');
+        }
         $informationCategories = InformationCategory::orderBy('name')->get();
         return view('admins.information.create-edit', compact(
             'information',
@@ -84,6 +97,9 @@ class InformationController extends Controller
      */
     public function update(InformationRequest $request, Information $information)
     {
+        if (!Auth::user()->can('Edit Informasi')) {
+            return redirect()->back()->with('error', 'Maaf, Anda tidak memiliki akses untuk halaman tersebut');
+        }
         $data = $request->validated();
         if ($request->hasFile('image')) {
             file_exists($information->image) ? unlink($information->image) : null;
@@ -99,6 +115,9 @@ class InformationController extends Controller
      */
     public function destroy(Information $information)
     {
+        if (!Auth::user()->can('Delete Informasi')) {
+            return redirect()->back()->with('error', 'Maaf, Anda tidak memiliki akses untuk halaman tersebut');
+        }
         file_exists($information->image) ? unlink($information->image) : null;
         $information->delete();
         return redirect()->route('information.index')
